==> path_3b/R/step-2_individual-data-reformatting.Rmd <==
---
title: "Path 3.B - Step 2 - Individual data reformatting"
date: "`r Sys.Date()`"
output: 
  html_document:
    theme: "cosmo"
    highlight: tango
    toc: true
    toc_float:
      collapsed: false
      smooth_scroll: false
    toc_depth: 4
---

```{r setup, include=FALSE}

knitr::opts_chunk$set(echo = TRUE)

```

# Load packages

```{r base}

# 1. Required packages ----

library(tidyverse) # Core tidyverse packages
library(readxl) # To read excel files
library(lubridate) # To transform date

# 2. Create a vector of selected variables names to check which ones are missing ----

selected_variables <- c("DatasetID", "Area", "Country", "Archipelago", 
                        "Location", "Site", "Replicate", "Zone", "Latitude", "Longitude", 
                        "Depth", "Year", "Date", "Method", "Observer", "Tax_ID", "Density", "Size", "Size_type")

```

# Data cleaning

## Data contributor 1

```{r DATA1}

# 1. Import data ----

# Main data

data_main <- read_excel("./../data/01_raw/data_contributor_1/dataset1.xlsx", sheet = "Data")

# Site data

data_site <- read_excel("./../data/01_raw/data_contributor_1/dataset1.xlsx", sheet = "Sites") %>% 
  rename(Observer = Diver) %>% 
  mutate(Latitude = as.numeric(as.character(Latitude)),
         Longitude = as.numeric(as.character(Longitude)))

# 2. Merge and modifications ----

data_main <- data_main %>% 
  select(-Group) %>% # Delete useless variables
  rename(Tax_ID = Species, Replicate = Transect, Size = Length) %>% # Rename variables
  group_by(Year, Date, Site, Replicate, Tax_ID, Size) %>% 
  summarise(Density = n()) %>% # Calculate the number of fish by species and length
  ungroup() %>%
  left_join(., data_site) %>% # Merge main data with sites data
  mutate(DatasetID = "DATA1",
         Area = "Caribbean",
         Country = "Venezuela",
         Location = "Chichiriviche",
         Method = "Belt transect",
         Size_type = "Total length") # Add variables absent in data (can be obtained through data contributor)

# 3. Export data ----

write.csv2(data_main, "./../data/02_reformatted/02-reformatted_DATA1.csv", row.names = FALSE)

# 4. Missing variables ----

setdiff(selected_variables, colnames(data_main)) # Get names of missing variables

# 5. Remove useless datasets and variables ----

rm(data_main, data_site)

```

## Data contributor 2

```{r}

# 1. Import data ----

# Main data

data_main <- read.csv2("./../data/01_raw/data_contributor_2/dataset2a.csv")

# Site data

data_site <- read.csv2("./../data/01_raw/data_contributor_2/dataset2b.csv", dec = ".")

# 2. Merge and modifications ----

data_main <- data_main %>% 
  pivot_longer("PGI12":"ISL14", names_to = "Name", values_to = "Density") %>% 
  filter(Density != 0, !(is.na(Density))) %>% 
  mutate(Year = paste0("20", str_sub(Name, 4, 5)),
         Name = str_sub(Name, 1, 3)) %>% 
  left_join(., data_site) %>% 
  select(-Name) %>% 
  rename(Tax_ID = Sp, Latitude = Lat, Longitude = Long) %>% 
  mutate(DatasetID = "DATA2",
         Area = "Caribbean",
         Country = "St. Lucia",
         Method = "Belt transect",
         Size_type = "Total length") # Add variables absent in data (can be obtained through data contributor)

# 3. Export data ----

write.csv2(data_main, "./../data/02_reformatted/02-reformatted_DATA2.csv", row.names = FALSE)

# 4. Missing variables ----

setdiff(selected_variables, colnames(data_main)) # Get names of missing variables

# 5. Remove useless datasets and variables ----

rm(data_main, data_site)

```

## Data contributor 3

```{r}

# 1. Import data ----

# Main data

data_main <- read_excel("./../data/01_raw/data_contributor_3/dataset3.xlsx", sheet = 1)

# Site data

data_site <- read_excel("./../data/01_raw/data_contributor_3/dataset3.xlsx", sheet = 2)

# Code data

data_code <- read_excel("./../data/01_raw/data_contributor_3/dataset3.xlsx", sheet = 3)

# 2. Merge and modifications ----

data_main <- data_main %>% 
  rename(Site_code = Site, Sp_code = Species) %>% # Rename variables for the join
  left_join(., data_site) %>% # Join main data with site data
  left_join(., data_code) %>% # Join main data with species code
  select(-Sp_code, -Site_code) %>% 
  rename(Replicate = Transect, Density = Number, Location = Island, 
         Tax_ID = Sp_name, Site = Site_name) %>% 
  mutate(DatasetID = "DATA3",
         Area = "Pacific",
         Density = Density/2, # Convert n.ind.200 m-2 to 100 m-2
         Archipelago = "Society Islands",
         Country = "France",
         Method = "Belt transect",
         Size_type = "Total length") # Add variables absent in data (can be obtained through data contributor)

# 3. Export data ----

write.csv2(data_main, "./../data/02_reformatted/02-reformatted_DATA3.csv", row.names = FALSE)

# 4. Missing variables ----

setdiff(selected_variables, colnames(data_main)) # Get names of missing variables

# 5. Remove useless datasets and variables ----

rm(data_main, data_site, data_code)

```

## Data contributor 4

```{r}

# 1. Import data ----

# Import first sheet

data_main <- read_excel("./../data/01_raw/data_contributor_4/dataset4a.xlsx", sheet = 1)
  
# Make a loop to bind all sheets

for(i in 2:length(excel_sheets("./../data/01_raw/data_contributor_4/dataset4a.xlsx"))){
  
  data_main <- read_excel("./../data/01_raw/data_contributor_4/dataset4a.xlsx", sheet = i) %>% 
    bind_rows(data_main, .)
  
}

# Site data

data_site <- read.csv2("./../data/01_raw/data_contributor_4/dataset4b.csv", dec = ".")

# 2. Merge and modifications ----

data_main <- data_main %>% 
  group_by(Site, StationName, Species, MeanSize) %>% 
  summarise(NbFishes = sum(NbFishes, na.rm = TRUE)) %>% # Make the sum of fish by species and length
  ungroup() %>% 
  left_join(., data_site) %>% # Join main data with site data
  rename(Replicate = StationName, Tax_ID = Species, Density = NbFishes, 
         Size = MeanSize, Location = Island, Depth = Depth_ft) %>% 
  mutate(Date = format(strptime(as.character(Date), "%d/%m/%Y"), "%Y-%m-%d"), # Change the date format
         Depth = Depth/3.281, # Convert depth from feet to meters
         Replicate = str_sub(Replicate, 3, 3)) %>% # Extract only the number of replicate 
  mutate(DatasetID = "DATA4",
         Area = "Pacific",
         Archipelago = "Solomon Islands",
         Method = "Belt transect") # Add variables absent in data (can be obtained through data contributor)

# 3. Export data ----

write.csv2(data_main, "./../data/02_reformatted/02-reformatted_DATA4.csv", row.names = FALSE)

# 4. Missing variables ----

setdiff(selected_variables, colnames(data_main)) # Get names of missing variables

# 5. Remove useless datasets and variables ----

rm(data_main, data_site)

```

## Data contributor 5

```{r}

# 1. Import data ----

# List of files containing main data (can be done automatically if the number of files is too important)

list_maindata <- c("dataset5a", "dataset5b", "dataset5c")

# Import first sheet

data_main <- read_excel(paste0("./../data/01_raw/data_contributor_5/", list_maindata[1], ".xlsx"), sheet = 1)

# Make a loop to bind all files

for(i in 2:length(list_maindata)){
  
  data_main <- read_excel(paste0("./../data/01_raw/data_contributor_5/", list_maindata[i], ".xlsx"), sheet = 1) %>% 
    bind_rows(data_main, .)
    
}

# Site data

data_site <- read_excel("./../data/01_raw/data_contributor_5/dataset5d.xlsx", sheet = 1) %>% 
  mutate(Latitude = as.numeric(as.character(Latitude)),
         Longitude = as.numeric(as.character(Longitude)))

# 2. Merge and modifications ----

data_main <- data_main %>% 
  rename(Code = Site, Tax_ID = Species, Density = Abundance, Size = Length) %>% 
  group_by(Code, Tax_ID, Size) %>% 
  summarise(Density = sum(Density, na.rm = TRUE)) %>% # Make the sum of fish by species and length
  ungroup() %>% 
  left_join(., data_site) %>% # Extract only the number of replicate 
  mutate(Replicate = str_sub(Code, 4, 4)) %>% 
  select(-Code) %>% 
  mutate(DatasetID = "DATA5",
         Area = "Pacific",
         Country = "Republic of Fiji",
         Archipelago = "Fiji Islands",
         Size_type = "Total length") # Add variables absent in data (can be obtained through data contributor)

# 3. Export data ----

write.csv2(data_main, "./../data/02_reformatted/02-reformatted_DATA5.csv", row.names = FALSE)

# 4. Missing variables ----

setdiff(selected_variables, colnames(data_main)) # Get names of missing variables

# 5. Remove useless datasets and variables ----

rm(data_main, data_site, i)

```

# Reproducibility

```{r reprod}

# 1. Reproducibility ----

sessionInfo()

```

---
Jeremy WICQUART | jeremywicquart@gmail.com | `r format(Sys.time())`
