==> path_3b/R/step-4_quality-assurance-quality-control.Rmd <==
---
title: "Path 3.B - Step 4 - Quality assurance and quality control"
date: "`r Sys.Date()`"
output: 
  html_document:
    theme: "cosmo"
    highlight: tango
    toc: true
    toc_float:
      collapsed: false
      smooth_scroll: false
    toc_depth: 4
---

```{r setup, include=FALSE}

knitr::opts_chunk$set(echo = TRUE)

```

# Load data and packages

```{r base}

# 1. Required packages ----

library(tidyverse) # Core tidyverse packages
library(formattable)
library(DT)
library(leaflet)

# 2. Import data ----

synthetic_data <- read.csv2("./../data/03_synthetic-dataset.csv")

```

# Missing variables

```{r}

# 1. Missing variables* by dataset ----
# * variables present but only filled with NA

synthetic_data %>% 
  group_by(DatasetID) %>% 
  summarise_all(~(all(is.na(.)))) %>%
  pivot_longer(2:ncol(.), names_to = "Variable", values_to = "State") %>% 
  filter(State == TRUE) %>% 
  select(-State) %>% 
  group_by(DatasetID) %>% 
  mutate(Variable = paste(Variable, collapse = ", ")) %>% 
  unique(.) %>% # Remove duplicates
  arrange(DatasetID) %>% 
  formattable(.) %>% 
  as.datatable(., rownames = FALSE)

```

# Check duplicates

```{r}

# 1. Check for duplicated datasets integrated ----

synthetic_data %>% 
  # Remove variables that may have been completed differently 
  # in the data reformatting of the datasets
  select(-DatasetID, -Area, -Country, -Archipelago, 
         -Location, -Zone, -Method, -Observer) %>% 
  group_by_all() %>% 
  count() %>% 
  filter(n > 1) %>% 
  formattable(.) %>% 
  as.datatable(., rownames = FALSE)

```

# Check errors

## Quali. variables

```{r}

# 1. List of levels of factor of (some) qualitative variables by datasetID ----

# 1.1 Vector of variables names to list --

quali_var <- c("Area", "Country", "Archipelago", "Location", "Site", "Zone", "Method")

# 1.2 Make a for loop to list levels of factor for each variable --

for (i in 1:length(quali_var)) {
    
  list_i <- synthetic_data %>% 
    select(quali_var[i]) %>% 
    unique()
  
  print(list_i)
  
}

```

## Quanti. variables

```{r}

# 1. Histogram of (some) quantitative variables by datasetID ----

# 1.1 Vector of variables names to plot --

quanti_var <- c("Depth", "Year", "Density", "Size")

# 1.2 Make a for loop to plot each variable --

for (i in 1:length(quanti_var)) {
    
  plot_i <- ggplot() +
    #geom_histogram(data = synthetic_data %>% select(-DatasetID), 
    #               aes_string(x = quanti_var[i]), color = "lightgrey", fill = "lightgrey") +
    geom_histogram(data = synthetic_data, aes_string(x = quanti_var[i]), 
                   color = "#16a085", fill = "#2abb9b") +
    theme_bw() +
    labs(title = quanti_var[i]) +
    facet_wrap(~DatasetID, ncol = 3)
    
  print(plot_i)
    
}

```

## Site coordinates

```{r fig.width=8}

# 1. Map of all sites locations ----

synthetic_data %>% 
  select(DatasetID, Longitude, Latitude) %>% 
  unique() %>% 
  leaflet(data = .) %>%
  addTiles() %>%
  addMarkers(~Longitude, ~Latitude, popup = ~as.character(DatasetID), 
             label = ~as.character(DatasetID))

```

# Export final file

```{r}

# 1. Remove rows with errors ----

synthetic_data_clean <- synthetic_data %>% 
  drop_na(Density) %>% # Remove NA values for Density
  filter(Density >= 0) %>%  # Remove negative values for Density
  filter(Size >= 0 | is.na(Size)) # Remove negative values for Size but keep NA

# 2. Number and percentage of rows removed by dataset ----

left_join(synthetic_data %>% 
            group_by(DatasetID) %>% 
            count(name = "Before"),
          inner_join(synthetic_data, synthetic_data_clean) %>% 
            group_by(DatasetID) %>% 
            count(name = "After")) %>% 
  mutate(Removed_abs = Before - After,
         Removed_perc = (Removed_abs/Before)*100) %>% 
  formattable(., list(Removed = color_bar("#e74c3c"))) %>% 
  as.datatable(., rownames = FALSE, colnames = c("Dataset", "n rows before", 
                                                 "n rows after", "n rows removed", 
                                                 "n rows removed (%)"))

```

```{r}

# 1. Export final file ----

synthetic_data_clean %>% 
  select(DatasetID, Area, Country, Archipelago, 
         Location, Site, Zone, Latitude, Longitude, 
         Depth, Year, Date, Replicate, Method, Observer, 
         Family, Genus, Species, Density, Size) %>% # Re-order variables
  write.csv2(., "./../data/04_final-synthetic-dataset.csv", row.names = FALSE) # Export the data

```

# Reproducibility

```{r reprod}

# 1. Reproducibility ----

sessionInfo()

```

---
Jeremy WICQUART | jeremywicquart@gmail.com | `r format(Sys.time())`
